final/requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1
twilio>=9.0.0
google-generativeai>=0.8.0
pydantic>=2.8.0
